move validate data to service

+++ services/post.service.js
@@ -6,20 +6,31 @@ const CommentReact = require("../models/comment_react.model");
 exports.getAllCmtDesc = async (id, sort) => {
   try {
     const ordered = [];
-
+    let message, comment;
     // query for sort comment by descend timestamp
     if (sort === "-created") {
       ordered.push(["created_at", "DESC"]);
     }
-    // find all comments of current post and sort comments by lateset timestamp
-    let comment = await Comment.findAll({
-      where: {
-        post_id: id,
-      },
-      // Order condition
-      order: ordered,
-    });
-    return comment;
+
+    let isPostExists = await checkPostExistence(id);
+    // Check if there is a post in database
+    if (!isPostExists) {
+      message = "Post does not exist!";
+      comment = null;
+      return { message, comment };
+    } else {
+      message = null;
+        // find all comments of current post and sort comments by lateset timestamp
+      comment = await Comment.findAll({
+        where: {
+          post_id: id,
+        },
+        // Order condition
+        order: ordered,
+      });
+      return { message, comment };
+    }
+
   } catch (err) {
     throw err;
   }
@@ -28,13 +39,29 @@ exports.getAllCmtDesc = async (id, sort) => {
 // Delete a comments of a post
 exports.deleteComment = async (post_id, comment_id) => {
   try {
-      // find a comment and delete
-    await Comment.destroy({
-      where: {
-        id: comment_id,
-        post_id: post_id,
-      },
-    });
+    let isPostExists = await checkPostExistence(post_id);
+    let isCommentExists = await checkCommentExistence(comment_id);
+    let message
+    // Check if there is a post in database
+    if (!isPostExists) {
+      message = "Post does not exist!";
+      return message;
+    } else {
+      // Check if there is a comment in database
+      if (!isCommentExists) {
+        message = "Comment does not exist!";
+        return message;
+      } else {
+        // find a comment and delete
+        await Comment.destroy({
+          where: {
+            id: comment_id,
+            post_id: post_id,
+          },
+        });
+        return null;
+      }
+    }
   } catch (err) {
     throw err;
   }
@@ -42,13 +69,24 @@ exports.deleteComment = async (post_id, comment_id) => {
 
 exports.likeComment = async (user_id, comment_id) => {
   try {
-    // Create a new like
-    let like = await CommentReact.create({
-      user_id,
-      comment_id,
-    });
-    return like;
+    let isCommentExists = await checkCommentExistence(comment_id);
+    let message, like;
+    // Check if there is a comment in database
+    if (!isCommentExists) {
+      message = "Comment does not exist!";
+      like = null;
+      return message;
+    } else {
+      message = "Liked!";
+      // Create a new like
+      like = await CommentReact.create({
+        user_id,
+        comment_id,
+      });
+      return message;
+    }
   } catch (err) { // Call API again with the same user_id and comment_id will cause error
+    let message = "Unliked!";
     // Destroy like when call twice
     let like = await CommentReact.destroy({
       where: {
@@ -56,13 +94,13 @@ exports.likeComment = async (user_id, comment_id) => {
         comment_id,
       },
     });
-    return like;
+    return message;
   }
 };
 
 
 // Functions check existence
-exports.checkPostExistence = async (id) => {
+const  checkPostExistence = async (id) => {
   //Check condition where the id exists
   try {
     if (!isNaN(id)) {
@@ -73,7 +111,7 @@ exports.checkPostExistence = async (id) => {
     return error;
   }
 };
-exports.checkCommentExistence = async (id) => {
+const checkCommentExistence = async (id) => {
   //Check condition where the id exists
   try {
     if (!isNaN(id)) {
